# BackEnd/swagger.yaml
{
  "openapi": "3.0.0",
  "info": {
    "title": "SecurCity API",
    "description": "Description of APIs provided by the application",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://{defaultHost}",
      "variables": {
        "defaultHost": {
          "default": "www.example.com"
        }
      }
    }
  ],
  "paths": {
...
    "/api/v1/cittadino/setPassword/:id": {
      "put": {
        "tags": [
          "/api/v1/cittadino/setPassword/:id"
        ],
        "summary": "Set password",
        "description": "Set the new password after the recover process started.",
        "operationId": "Setpassword",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {
              "Authorization": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Bearer 4Ve6eAGcVS4q2q4UeNtDKt8hwKcjCXJ85W"
                  }
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Setpasswordresponse"
                    },
                    {
                      "example": {
                        "message": "You can now go back to the mobile app!"
                      }
                    }
                  ]
                },
                "example": {
                  "message": "You can now go back to the mobile app!"
                }
              }
            }
          },
          "404": {
            "description": "",
            "headers": {
              "Authorization": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Bearer 4Ve6eAGcVS4q2q4UeNtDKt8hwKcjCXJ85W"
                  }
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Setpasswordresponse1"
                    },
                    {
                      "example": {
                        "error": "User not found"
                      }
                    }
                  ]
                },
                "example": {
                  "error": "User not found"
                }
              }
            }
          },
          "500": {
            "description": "",
            "headers": {
              "Authorization": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Bearer 4Ve6eAGcVS4q2q4UeNtDKt8hwKcjCXJ85W"
                  }
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Setpasswordresponse1"
                    },
                    {
                      "example": {
                        "error": "Internal Server Error"
                      }
                    }
                  ]
                },
                "example": {
                  "error": "Internal Server Error"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Setpasswordresponse": {
        "title": "Setpasswordresponse",
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "You can now go back to the mobile app!"
          }
        }
      },
      "Setpasswordresponse1": {
        "title": "Setpasswordresponse1",
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "User not found"
          }
        }
      },
...
    },
    "securitySchemes": {
      "httpBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "httpBearer": []
    }
  ],
...
}